Refactor use yargs sub commands
