+++ src/cli/tool_commands/prompt/builder.ts
@@ -1,0 +1,14 @@
+import { Argv } from "yargs";
+import { getAvailableTemplates } from "@/src/utils/file.js";
+
+const availableTemplateNames = getAvailableTemplates().map((template) => template.filename);
+
+export const builder = (yargs: Argv) => {
+  return yargs.option("t", {
+    alias: "template",
+    describe: "Template name to use",
+    demandOption: false,
+    choices: availableTemplateNames,
+    type: "string",
+  });
+};
