+++ src/cli/tool_commands/prompt/index.ts
@@ -1,0 +1,4 @@
+export const command = "prompt";
+export const desc = "Dump prompt from template";
+export { builder } from "./builder.js";
+export { handler } from "./handler.js";
