+++ src/cli/tool_commands/schema/builder.ts
@@ -1,0 +1,5 @@
+import { Argv } from "yargs";
+
+export const builder = (yargs: Argv) => {
+  return yargs;
+};
