+++ src/cli/tool_commands/schema/index.ts
@@ -1,0 +1,4 @@
+export const command = "schema";
+export const desc = "Dump mulmocast schema";
+export { builder } from "./builder.js";
+export { handler } from "./handler.js";
