+++ src/cli/tool_commands/scripting/builder.ts
@@ -1,0 +1,74 @@
+import { Argv } from "yargs";
+import { languages } from "../../../utils/const.js";
+import { llmAgents } from "../../../utils/utils.js";
+import { getAvailableTemplates } from "../../../utils/file.js";
+
+const availableTemplateNames = getAvailableTemplates().map((template) => template.filename);
+
+export const builder = (yargs: Argv) => {
+  return yargs
+    .option("o", {
+      alias: "outdir",
+      description: "output dir",
+      demandOption: false,
+      type: "string",
+    })
+    .option("b", {
+      alias: "basedir",
+      description: "base dir",
+      demandOption: false,
+      type: "string",
+    })
+    .option("l", {
+      alias: "lang",
+      description: "target language",
+      choices: languages,
+      demandOption: false,
+      type: "string",
+    })
+    .option("u", {
+      alias: "url",
+      description: "URLs to reference (required when not in interactive mode)",
+      demandOption: false,
+      default: [],
+      type: "array",
+      string: true,
+    })
+    .option("i", {
+      alias: "interactive",
+      description: "Generate script in interactive mode with user prompts",
+      demandOption: false,
+      type: "boolean",
+    })
+    .option("t", {
+      alias: "template",
+      description: "Template name to use",
+      demandOption: false,
+      choices: availableTemplateNames,
+      type: "string",
+    })
+    .option("c", {
+      alias: "cache",
+      description: "cache dir",
+      demandOption: false,
+      type: "string",
+    })
+    .option("s", {
+      alias: "script",
+      description: "script filename",
+      demandOption: false,
+      default: "script",
+      type: "string",
+    })
+    .option("llm_agent", {
+      description: "llm agent",
+      demandOption: false,
+      choices: llmAgents,
+      type: "string",
+    })
+    .option("llm_model", {
+      description: "llm model",
+      demandOption: false,
+      type: "string",
+    });
+};
