+++ src/cli/tool_commands/scripting/index.ts
@@ -1,0 +1,4 @@
+export const command = "scripting";
+export const desc = "Generate mulmocast script";
+export { builder } from "./builder.js";
+export { handler } from "./handler.js";
