+++ src/types/cli_types.ts
@@ -1,0 +1,10 @@
+import { Arguments } from "yargs";
+
+export type GlobalOptions = {
+  v?: boolean;
+  o?: string;
+  b?: string;
+  l?: string;
+};
+
+export type ToolCliArgs<T> = Arguments<T & GlobalOptions>;
